updated readme and structure, add collection and refactoring
