+++ Product - Boticario.postman_collection.json
@@ -1,0 +1,106 @@
+{
+	"info": {
+		"_postman_id": "b17672af-8b36-48c9-9f95-2cc79df9a93d",
+		"name": "Product - Boticario",
+		"schema": "https://schema.getpostman.com/json/collection/v2.1.0/collection.json"
+	},
+	"item": [
+		{
+			"name": "Create Product",
+			"request": {
+				"method": "POST",
+				"header": [],
+				"body": {
+					"mode": "raw",
+					"raw": "{\r\n  \"sku\": 43264,\r\n  \"name\": \"L'Oréal Professionnel Expert Absolut Repair Cortex Lipidium - Máscara de Reconstrução 500g\",\r\n  \"inventory\": {\r\n    \"warehouses\": [\r\n      {\r\n        \"locality\": \"SP\",\r\n        \"quantity\": 22,\r\n        \"type\": \"ECOMMERCE\"\r\n      },\r\n      {\r\n        \"locality\": \"MOEMA\",\r\n        \"quantity\": 18,\r\n        \"type\": \"PHYSICAL_STORE\"\r\n      }\r\n    ]\r\n  }\r\n}",
+					"options": {
+						"raw": {
+							"language": "json"
+						}
+					}
+				},
+				"url": {
+					"raw": "http://localhost:8080/products",
+					"protocol": "http",
+					"host": [
+						"localhost"
+					],
+					"port": "8080",
+					"path": [
+						"products"
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "Get Product",
+			"request": {
+				"method": "GET",
+				"header": [],
+				"url": {
+					"raw": "http://localhost:8080/products/43264",
+					"protocol": "http",
+					"host": [
+						"localhost"
+					],
+					"port": "8080",
+					"path": [
+						"products",
+						"43264"
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "Update Product",
+			"request": {
+				"method": "PUT",
+				"header": [],
+				"body": {
+					"mode": "raw",
+					"raw": "{\r\n  \"name\": \"L'Oréal Professionnel Expert Absolut Repair Cortex Lipidium - Máscara de Reconstrução 500g\",\r\n  \"inventory\": {\r\n    \"warehouses\": [\r\n      {\r\n        \"locality\": \"SP\",\r\n        \"quantity\": 26,\r\n        \"type\": \"ECOMMERCE\"\r\n      },\r\n      {\r\n        \"locality\": \"PR\",\r\n        \"quantity\": 24,\r\n        \"type\": \"PHYSICAL_STORE\"\r\n      }\r\n    ]\r\n  }\r\n}",
+					"options": {
+						"raw": {
+							"language": "json"
+						}
+					}
+				},
+				"url": {
+					"raw": "http://localhost:8080/products/43264",
+					"protocol": "http",
+					"host": [
+						"localhost"
+					],
+					"port": "8080",
+					"path": [
+						"products",
+						"43264"
+					]
+				}
+			},
+			"response": []
+		},
+		{
+			"name": "Delete Product",
+			"request": {
+				"method": "DELETE",
+				"header": [],
+				"url": {
+					"raw": "http://localhost:8080/products/43264",
+					"protocol": "http",
+					"host": [
+						"localhost"
+					],
+					"port": "8080",
+					"path": [
+						"products",
+						"43264"
+					]
+				}
+			},
+			"response": []
+		}
+	]
+}
