+++ README.md
@@ -1,74 +1,66 @@
-### Backend Test
-[![Build Status](https://travis-ci.org/belezanaweb/test-nodejs.svg?branch=master)](https://travis-ci.org/belezanaweb/test-nodejs)
+### test-nodejs Boticario
 
-Esta é uma avaliação básica de código.
+## Requirements
+- Node (v18.16.0)
+- NPM (v9.5.1) or yarn (v1.22.19)
 
-O objetivo é conhecer um pouco do seu conhecimento/prática de RESTful e NodeJS.
+## Install dependencies
+```sh
+$ npm i
+```
+## Running project
+```sh
+$ npm run start:dev
+```
 
-Recomendamos que você não gaste mais do que 4 - 6 horas.
+## Running tests
+```sh
+$ npm run test
+```
 
-Faça um fork deste repositório.
+## Documentation
 
-Ao finalizar o teste, submeta um pull request para o repositório que nosso time será notificado.
+- Apos rodar o projeto acesse o link `http://localhost:8080/api`
+
+## UseCases
+- Create Product
+- Get Product
+- Update Product
+- Delete Product
 
 ### Tarefas
 
-Com a seguinte representação de produto:
-
-```json
-{
-    "sku": 43264,
-    "name": "L'Oréal Professionnel Expert Absolut Repair Cortex Lipidium - Máscara de Reconstrução 500g",
-    "inventory": {
-        "quantity": 15,
-        "warehouses": [
-            {
-                "locality": "SP",
-                "quantity": 12,
-                "type": "ECOMMERCE"
-            },
-            {
-                "locality": "MOEMA",
-                "quantity": 3,
-                "type": "PHYSICAL_STORE"
-            }
-        ]
-    },
-    "isMarketable": true
-}
-```
-
 Crie endpoints para as seguintes ações:
 
-- [ ] Criação de produto onde o payload será o json informado acima (exceto as propriedades **isMarketable** e **inventory.quantity**)
+- [x] Criação de produto onde o payload será o json informado acima (exceto as propriedades **isMarketable** e **inventory.quantity**)
 
-- [ ] Edição de produto por **sku**
+- [x] Edição de produto por **sku**
 
-- [ ] Recuperação de produto por **sku**
+- [x] Recuperação de produto por **sku**
 
-- [ ] Deleção de produto por **sku**
+- [x] Deleção de produto por **sku**
 
 ### Requisitos
 
 
-- [ ] Toda vez que um produto for recuperado por **sku** deverá ser calculado a propriedade: **inventory.quantity**
+- [x] Toda vez que um produto for recuperado por **sku** deverá ser calculado a propriedade: **inventory.quantity**
 
         A propriedade inventory.quantity é a soma da quantity dos warehouses
 
-- [ ] Toda vez que um produto for recuperado por **sku** deverá ser calculado a propriedade: **isMarketable**
+- [x] Toda vez que um produto for recuperado por **sku** deverá ser calculado a propriedade: **isMarketable**
 
         Um produto é marketable sempre que seu inventory.quantity for maior que 0
 
-- [ ] Caso um produto já existente em memória tente ser criado com o mesmo **sku** uma exceção deverá ser lançada
+- [x] Caso um produto já existente em memória tente ser criado com o mesmo **sku** uma exceção deverá ser lançada
 
         Dois produtos são considerados iguais se os seus skus forem iguais
 
 
-- [ ] Ao atualizar um produto, o antigo deve ser sobrescrito com o que esta sendo enviado na requisição
+- [x] Ao atualizar um produto, o antigo deve ser sobrescrito com o que esta sendo enviado na requisição
 
         A requisição deve receber o sku e atualizar com o produto que tbm esta vindo na requisição
 
-### Dicas
+### Observations
 
-- Os produtos podem ficar em memória, não é necessário persistir os dados
-- Testes são sempre bem-vindos :smiley:
+- Realizei a deleção para deleção logica (Soft-Delete)
+- Foi disponibilizado na raiz do projeto uma collection do postman de produtos `Product - Boticario.postman_collection.json`
