+++ src/app.ts
@@ -1,8 +1,8 @@
 import "reflect-metadata";
 import express, { Request, Response, NextFunction } from "express";
 import "express-async-errors"
-import swaggerUi  from 'swagger-ui-express';
 import router from "./routes";
+import swaggerUi  from 'swagger-ui-express';
 import swaggerDocs from './swagger.json';
 import { AppError } from "./shared/excepetions/errors";
 import '../src/shared/container'
