+++ src/domain/entities/products/Product.spec.ts
