+++ src/domain/entities/products/Product.ts
