+++ src/domain/entities/products/interface/Inventory.ts
