+++ src/domain/repositories/IProducts.repository.ts
@@ -1,6 +1,6 @@
 import { IProduct } from "../entities/products/Product"
-import { ICreateProductDTO } from "../modules/products/dtos/ICreateProductDTO"
-import { IUpdateProductDTO } from "../modules/products/dtos/IUpdateProductDTO"
+import { ICreateProductDTO } from "../../modules/products/dtos/ICreateProductDTO"
+import { IUpdateProductDTO } from "../../modules/products/dtos/IUpdateProductDTO"
 
 interface IProductsRepository {
   create(product: ICreateProductDTO): Promise<IProduct>
