+++ src/modules/products/service/IStockProduct.service.ts
@@ -1,4 +1,4 @@
-import { IProduct } from "../../../entities/products/Product";
+import { IProduct } from "../../../domain/entities/products/Product";
 
 interface IStockProductService {
   execute(product: IProduct): Promise<IProduct>
