+++ src/modules/products/service/StockProduct.service.ts
@@ -1,4 +1,4 @@
-import { IProduct } from "../../../entities/products/Product";
+import { IProduct } from "../../../domain/entities/products/Product";
 
 export class StockProductService {
   async execute (product: IProduct): Promise<IProduct> {
