+++ src/server.ts
@@ -4,6 +4,7 @@ import { app } from "./app"
 import mongoose from 'mongoose'
 
 const PORT = process.env.PORT
+const HOST = process.env.HOST_NAME
 const password = process.env.PASSWORD_MONGO_ATLAS
 const user = process.env.USER_MONGO_ATLAS
 
@@ -11,4 +12,4 @@ mongoose.connect(`mongodb+srv://${user}:${password}@boticario.9b0wn6v.mongodb.ne
   .then(() => console.log("Database Connected"))
   .catch((err) => { throw err })
 
-app.listen(PORT, () => console.log(`Magic happen on ${process.env.HOST_NAME}:${PORT}`))
+app.listen(PORT, () => console.log(`Server is running on ${HOST}:${PORT}`))
