+++ src/shared/container/repositoriesContainer.ts
@@ -1,5 +1,5 @@
 import { container } from "tsyringe"
-import { IProductsRepository } from "../../repostiories/IProducts.repository"
-import { MongoRepository } from "../../repostiories/mongoDB/Mongo.repository"
+import { IProductsRepository } from "../../domain/repositories/IProducts.repository"
+import { MongoRepository } from "../../domain/repositories/Mongo.repository"
 
 container.registerSingleton<IProductsRepository>("ProductsRepository", MongoRepository)
